Major refactor of the properties module!
Due to a major bug in the updateProperty controller, I decided to create
this branch to fix that bug specifically. However, I ended up building a
mini-app to help me understand JSON in MsSQL better, and this ended up
in me rewritting the entire properties module.

The addProperty,getProperties,getPropertyById,deleteProperty were
basically untouched; save for a little bit of reshuffling aroung the
storedprocedures. That has allowed me to manipulate JSON better.

In the updateUser, I removed the JSON_MODIFY() function that might have
been leading up to the multiple escape characters within the sequence.
Using instead the values fetched as is from the request, and modified
slightly with typescripts Json.stringify() function.

A major win in this PR was getting to understand helpers better. I built
the properties search helper, that has allowed one to seach for a
specific property based on their specified inputs. There is no
storedProcedure for this yet but instead everything is built dynamically
in the controller. This creates modularity that allows for easier code
reusability.

Turns out JSON in SQL isnt so bad after all!