*NOT* CLOSED ON SUNDAY(MAJOR BUG)
The main issue here is nesting. In trying to add properties with nested
values, alot of error have arisen.

- The first was validation using joi schema. This was promptly resolved
  by utilizing nesting in the validation svhema, by initializing the
keys to objects, then the nested values to the respective values, either
string or number.
- Then came at the database level. I needed to find a way to add JSON
  structured data into my SQL database. This turned out to be quite easy
due to the introduction of JSON in MsSQL. Proper documentation was also
available.
- Being able to quickly add and modify JSON columns on the raw SQL, the
  major issue came in when integrating this at th eapplication level. In
the addProperty controller to be specific.
- I took it a step down, removed the database helpers and tried
  implementing this using bare SQL queries with the help of the mssql
package, yet I got the same error. EPARAM
- This error seems to stem from the database side
- Need to look further into it. Maybe write a blog on this