ERROR RESOLVED
Resolved the major error that prevented me from adding properties into
the system. This was after having a better undertanding of the JSON
datatype in MSSQL and also the JSON methods in javascript.

Steps followed included:
- Jsonify the details and location values once extracted from the
  request body.
- Feed these Jsonfied values into the stored procedure instead of the
  original values.
- In the SQL tables, values with JSON data have data type NVARCHAR(MAX)

This proved to be an important process as by mastering this I understood
how to imput nested json values into my SQL db, alongside normalized
columns. This could prove to be quite essential, case in point:
citizenConnect 360.

In the polls table, instead of adding choices as a string of CSV's, One
could instead input them as JSON with key values pairs.