ERROR ON UPDATEPROPERTY CONTROLLER
While building the getPropertyBy controller, I realized some bugs in the
updateProperty Controller:
-  While updating the property, more and more escape characters were being added to the values within the databse. Most likely dues to the JSON.stringify() function within the controller.
- In trying to eliminate this abnormally by using the JSON_QUERY() function to properly format JSON, it seemed to append itself on the existing values. Removing the escape characters error and adding a new one.
- This has lead me to rethink on the neccessity of using JSON schema
  within my database overall.

For the getPropertyBy controller, I need to research on usage of utils
in Node.Js projects. With this understanding, I intend to build a simple
utils to identify the search parameters being passed.

An alternative mastering the use of search parameters. When a user
selects those, they are passed to the specified url using the next()
function and so on....